+++ governance/xc_admin/packages/xc_admin_common/src/programs/program_registry.ts
@@ -3,11 +3,8 @@ import { PythCluster } from "@pythnetwork/client";
 import {
   DownloadableConfig,
   ProgramType,
-  RawConfig,
-  LazerConfig,
   ProgramConfig,
   GetConfigParams,
-  LazerConfigParams,
   ValidationResult,
 } from "./types";
